Added int64 type to database types. Moved testlogger into testutil package

+++ model/jobs.go
@@ -3,7 +3,7 @@ package model
 // SubmitJob represents an incoming request to persist a job.
 type SubmitJob struct {
 	Name      string `json:"name"`
-	StartTime int    `json:"start_time"`
-	Interval  int    `json:"interval"`
+	StartTime int64  `json:"start_time"`
+	Interval  int64  `json:"interval"`
 	URL       string `json:"url"`
 }
